bug fix: variables to store CborError should be at least 32-bit width

+++ applets/ctap/ctap-parser.c
@@ -160,7 +160,7 @@ uint8_t parse_credential_descriptor(CborValue *arr, uint8_t *id) {
   if (cbor_value_get_type(arr) != CborMapType) return CTAP2_ERR_CBOR_UNEXPECTED_TYPE;
 
   CborValue val;
-  uint8_t ret = cbor_value_map_find_value(arr, "id", &val);
+  int ret = cbor_value_map_find_value(arr, "id", &val);
   CHECK_CBOR_RET(ret);
   if (cbor_value_get_type(&val) != CborByteStringType) return CTAP2_ERR_MISSING_PARAMETER;
   size_t len = sizeof(CredentialId);
